+++ src/simulation/structs.jl
@@ -74,6 +74,8 @@ mutable struct Host
     id::Int64
 
     sequence::String
+    mean_mutations_per_replication::Float64
+    mean_recombination_crossovers::Float64
 
     pathogens::Vector{Pathogen} # size MAX_PATHOGENS
     responses::Vector{Response} # size MAX_RESPONSES
@@ -102,6 +104,16 @@ struct PopulationType
     inoculum_coefficient::Float64
     mutation_coefficient::Float64
     recombination_coefficient::Float64
+
+    host_num_loci::Int64
+    host_possible_alleles::String
+    host_mean_mutations_per_replication::Float64
+    host_sexual_reproduction::Bool
+    host_mean_recombination_crossovers::Float64
+
+    hostSexualCompatibility::FunctionWrapper{Bool,Tuple{String,String}}
+    hostMutationCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
+    hostRecombinationCoefficient::FunctionWrapper{Float64,Tuple{String}} # takes seq argument, returns Float64
 
     base_coefficients::SVector{NUM_COEFFICIENTS,Float64}